+++ g_Qfmt_p.c
@@ -31,7 +31,7 @@ THIS SOFTWARE.
 
 #include "gdtoaimp.h"
 
- extern ULong NanDflt_Q_D2A[4];
+ extern uint32_t NanDflt_Q_D2A[4];
 
 #undef _0
 #undef _1
@@ -60,7 +60,7 @@ g_Qfmt_p(char *buf, void *V, int ndig, size_t bufsize, int nik)
 {
 	static FPI fpi0 = { 113, 1-16383-113+1, 32766 - 16383 - 113 + 1, 1, 0, Int_max };
 	char *b, *s, *se;
-	ULong bits[4], *L, sign;
+	uint32_t bits[4], *L, sign;
 	int decpt, ex, i, mode;
 #ifdef Honor_FLT_ROUNDS
 #include "gdtoa_fltrnds.h"
@@ -73,7 +73,7 @@ g_Qfmt_p(char *buf, void *V, int ndig, size_t bufsize, int nik)
 	if (bufsize < (size_t)(ndig + 10))
 		return 0;
 
-	L = (ULong*)V;
+	L = (uint32_t*)V;
 	sign = L[_0] & 0x80000000L;
 	bits[3] = L[_0] & 0xffff;
 	bits[2] = L[_1];
