+++ g_dfmt.c
@@ -40,7 +40,7 @@ g_dfmt(char *buf, double *d, int ndig, size_t bufsize)
 {
 	static FPI fpi0 = { 53, 1-1023-53+1, 2046-1023-53+1, 1, 0, Int_max };
 	char *b, *s, *se;
-	ULong bits[2], *L, sign;
+	uint32_t bits[2], *L, sign;
 	int decpt, ex, i, mode;
 #ifdef Honor_FLT_ROUNDS
 #include "gdtoa_fltrnds.h"
@@ -53,7 +53,7 @@ g_dfmt(char *buf, double *d, int ndig, size_t bufsize)
 	if (bufsize < (size_t)(ndig + 10))
 		return 0;
 
-	L = (ULong*)d;
+	L = (uint32_t*)d;
 	sign = L[_0] & 0x80000000L;
 	if ((L[_0] & 0x7ff00000) == 0x7ff00000) {
 		/* Infinity or NaN */
