+++ g_xfmt.c
@@ -60,8 +60,8 @@ g_xfmt(char *buf, void *V, int ndig, size_t bufsize)
 {
 	static FPI fpi0 = { 64, 1-16383-64+1, 32766 - 16383 - 64 + 1, 1, 0, Int_max };
 	char *b, *s, *se;
-	ULong bits[2], sign;
-	UShort *L;
+	uint32_t bits[2], sign;
+	uint16_t *L;
 	int decpt, ex, i, mode;
 #ifdef Honor_FLT_ROUNDS
 #include "gdtoa_fltrnds.h"
@@ -74,7 +74,7 @@ g_xfmt(char *buf, void *V, int ndig, size_t bufsize)
 	if (bufsize < (size_t)(ndig + 10))
 		return 0;
 
-	L = (UShort *)V;
+	L = (uint16_t *)V;
 	sign = L[_0] & 0x8000;
 	bits[1] = (L[_1] << 16) | L[_2];
 	bits[0] = (L[_3] << 16) | L[_4];
